refactor: project archive card
