+++ frontend/src/component/project/NewProjectCard/ProjectCardFooter/ProjectCardFooter.tsx
@@ -1,10 +1,10 @@
 import type React from 'react';
 import type { FC } from 'react';
 import { Box, styled } from '@mui/material';
-import { FavoriteIconButton } from 'component/common/FavoriteIconButton/FavoriteIconButton';
-import useToast from 'hooks/useToast';
-import { useFavoriteProjectsApi } from 'hooks/api/actions/useFavoriteProjectsApi/useFavoriteProjectsApi';
-import useProjects from 'hooks/api/getters/useProjects/useProjects';
+import {
+    type IProjectOwnersProps,
+    ProjectOwners,
+} from '../ProjectOwners/ProjectOwners';
 
 interface IProjectCardFooterProps {
     id: string;
@@ -12,6 +12,7 @@ interface IProjectCardFooterProps {
     children?: React.ReactNode;
     Actions?: FC<{ id: string; isFavorite?: boolean }>;
     disabled?: boolean;
+    owners: IProjectOwnersProps['owners'];
 }
 
 const StyledFooter = styled(Box)<{ disabled: boolean }>(
@@ -27,58 +28,15 @@ const StyledFooter = styled(Box)<{ disabled: boolean }>(
     }),
 );
 
-const StyledContainer = styled(Box)(({ theme }) => ({
-    padding: theme.spacing(1.5, 0, 2.5, 3),
-    display: 'flex',
-    alignItems: 'center',
-}));
-
-const StyledFavoriteIconButton = styled(FavoriteIconButton)(({ theme }) => ({
-    margin: theme.spacing(1, 2, 0, 0),
-}));
-
-const FavoriteAction: FC<{ id: string; isFavorite?: boolean }> = ({
-    id,
-    isFavorite,
-}) => {
-    const { setToastApiError } = useToast();
-    const { favorite, unfavorite } = useFavoriteProjectsApi();
-    const { refetch } = useProjects();
-
-    const onFavorite = async (e: React.SyntheticEvent) => {
-        e.preventDefault();
-        try {
-            if (isFavorite) {
-                await unfavorite(id);
-            } else {
-                await favorite(id);
-            }
-            refetch();
-        } catch (error) {
-            setToastApiError('Something went wrong, could not update favorite');
-        }
-    };
-
-    return (
-        <StyledFavoriteIconButton
-            onClick={onFavorite}
-            isFavorite={Boolean(isFavorite)}
-            size='medium'
-        />
-    );
-};
-
 export const ProjectCardFooter: FC<IProjectCardFooterProps> = ({
     children,
-    id,
-    isFavorite = false,
-    Actions = FavoriteAction,
+    owners,
     disabled = false,
 }) => {
     return (
         <StyledFooter disabled={disabled}>
-            <StyledContainer>{children}</StyledContainer>
-            <Actions id={id} isFavorite={isFavorite} />
+            <ProjectOwners owners={owners} />
+            {children}
         </StyledFooter>
     );
 };
